Remove wrapper/runlist combination
Following the same pattern as stack_commons, removing some of the duplicated tests we had. I verified each test was identical across default, runlist, and wrapper. I left only a default test suite which uses the wrapper test fixture to call it. This is the standard use case for the stacks, so this makes sense to be the one we keep (our use case is NOT to converge the stack directly or via runlist).